+++ internal/update/pubkey.go
@@ -8,4 +8,5 @@ package update
 // verificar firmas nuevas: reinstalar por npm.
 var releasePubKeys = []string{
 	"b1fe45a0ad9a21eb4f6d9e54f0b2806070cd3bf20cb94279a7cf735cb9012153",
+	"96c3f40308cefc60591b25ddc1615c6a9d1120ea14fa571eb3a0d97d9baf5abb",
 }